refactor: move handshake encryption to crypto package
