+++ lib/src/handshake/handshake.dart
@@ -1,7 +1,7 @@
 import 'dart:typed_data';
 
 import 'package:cryptography/cryptography.dart';
-import 'package:endguard/src/handshake/encryption.dart';
+import 'package:endguard/src/crypto/handshake_encryption.dart';
 import 'package:endguard/src/handshake/message.dart';
 import 'package:endguard/src/protos/protocol.pb.dart';
 import 'package:endguard/src/crypto/diffie_hellman_ratchet.dart';
